# frontend/public/demo-api/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryAudit": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "matched_rule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Rule"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "previous_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Category"
          },
          "previous_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Source"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          }
        },
        "title": "CategoryAudit",
        "type": "object"
      },
...
      "ExtractionAudit": {
        "description": "Provenance for a transaction recovered via the AI extraction fallback.\n\nPresent only on rows the parsers couldn't read but the constrained LLM\nextraction recovered (parse-failure-quarantine path). Absent on\nparser-extracted and statement-imported rows.",
        "properties": {
          "extracted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted At"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "validated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated"
          }
        },
        "title": "ExtractionAudit",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JournalDay": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "day_total": {
            "title": "Day Total",
            "type": "number"
          },
          "mtd_total": {
            "title": "Mtd Total",
            "type": "number"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "date",
          "day_total",
          "count",
          "mtd_total",
          "transactions"
        ],
        "title": "JournalDay",
        "type": "object"
      },
      "JournalResponse": {
        "properties": {
          "budget_ceiling": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Ceiling"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/JournalDay"
            },
            "title": "Days",
            "type": "array"
          },
          "month": {
            "title": "Month",
            "type": "string"
          },
          "month_total": {
            "title": "Month Total",
            "type": "number"
          },
          "transaction_count": {
            "title": "Transaction Count",
            "type": "integer"
          }
        },
        "required": [
          "month",
          "days",
          "month_total",
          "transaction_count",
          "budget_ceiling"
        ],
        "title": "JournalResponse",
        "type": "object"
      },
...
      "TransactionContext": {
        "properties": {
          "category_budget_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Budget Pct"
          },
          "category_budget_target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Budget Target"
          },
          "category_month_total": {
            "title": "Category Month Total",
            "type": "number"
          },
          "merchant_month_count": {
            "title": "Merchant Month Count",
            "type": "integer"
          }
        },
        "required": [
          "category_month_total",
          "merchant_month_count",
          "category_budget_target",
          "category_budget_pct"
        ],
        "title": "TransactionContext",
        "type": "object"
      },
...
      "TransactionResponse": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "category_audit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryAudit"
              },
              {
                "type": "null"
              }
            ]
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "date_file_name": {
            "title": "Date File Name",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "extraction_audit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionAudit"
              },
              {
                "type": "null"
              }
            ]
          },
          "forwarded_to": {
            "title": "Forwarded To",
            "type": "string"
          },
          "ignored": {
            "title": "Ignored",
            "type": "boolean"
          },
          "institution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "statement_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement Source"
          },
          "transaction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Type"
          },
          "tx_id": {
            "title": "Tx Id",
            "type": "string"
          }
        },
        "required": [
          "tx_id",
          "forwarded_to",
          "date_file_name",
          "date",
          "amount",
          "company",
          "category",
          "institution",
          "transaction_type",
          "name",
          "ignored",
          "comment",
          "deleted_at"
        ],
        "title": "TransactionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal finance dashboard API. Ingests bank transaction emails, categorizes them (with optional OpenAI enrichment), and exposes transactions, budgets, summaries, and AI spending insights over a consistent JSON HTTP interface.\n\n**Storage:** dual-backend \u2014 DynamoDB (AWS) or SQLite (local/demo). Selection is driven by `data/config.json`. All endpoints return identical shapes regardless of backend.\n\n**Versioning:** routes are prefixed with `/api/v1/`. Breaking changes will ship as `/api/v2/` rather than in-place.\n\n**Demo:** this is a read-only snapshot of the Tidings demo journal \u2014 the fictional Mira Lin Chen household, covering 2025-05 through 2026-03. All data is invented. Any request other than GET, HEAD, or OPTIONS returns 405; self-host Tidings for a writable journal with bearer-token auth.",
    "title": "Tidings demo API (read-only)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/journal": {
      "get": {
        "operationId": "getJournal",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "title": "Month",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Day-grouped transaction timeline for a month",
        "tags": [
          "journal"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://gettidings.com/demo"
    }
  ]
}